// src/app/models/entity-type.ts
export enum ScenarioType {
  USER_ACTION = 1, // Сценарий активируется по требованию пользователя
  SENSOR_ACTION = 2, // Сценарий активируется при определенных показаниях датчика
  SCHEDULE_ACTION = 3 // Сценарий активируется по расписанию
}

export enum ControllerType {
  CONTROLLER_TYPE_1 = 1,
  CONTROLLER_TYPE_2 = 2,
  CONTROLLER_TYPE_3 = 3,
  CONTROLLER_TYPE_4 = 4,
}

export enum SensorType {
  SENSOR_TYPE_1 = 1,
  SENSOR_TYPE_2 = 2,
  SENSOR_TYPE_3 = 3,
  SENSOR_TYPE_4 = 4,
}

export enum ThingType {
  THING_TYPE_1 = 1,
  THING_TYPE_2 = 2,
  THING_TYPE_3 = 3,
  THING_TYPE_4 = 4,
}